Refactor API
Moved API files into their own package.
Note: did not begin propagating Model name changes yet
